Identify `TransitionValidator` as part of the model. Extract to separate file. Update testing accordingly.
