+++ src/domain/services/update_order_status_service.py
@@ -1,7 +1,6 @@
 from domain.models.identifier import Identifier
 from domain.ports.api.update_order_status_api import (
     UpdateOrderStatusAPI,
-    ExpectednessSetting,
 )
 from domain.ports.spi.order_persistence_spi import UpdateOrderSPI, GetOrderByOrderIdSPI
 from domain.ports.spi.status_update_event_dispatcher_spi import (
@@ -11,7 +10,11 @@ from domain.models.order import PersistedOrder
 from domain.models.order_status import (
     Status,
     StatusTransition,
-    StatusTransitionProtocol,
+)
+from domain.models.status_transition_validator import (
+    ExpectednessSetting,
+    TransitionValidatorProtocol,
+    TransitionValidator,
 )
 from domain.models.event import (
     StatusToEventMapper,
@@ -19,39 +22,6 @@ from domain.models.event import (
     DispatchableEvent,
 )
 from domain.errors import InvalidOrderIdError, InsufficientExpectednessError
-from domain.utils.singleton_meta import SingletonMeta
-from typing import Protocol
-
-
-class TransitionValidatorProtocol(Protocol):
-    @classmethod
-    def validate_transition(
-        cls, transition: StatusTransitionProtocol, setting: ExpectednessSetting
-    ) -> bool:
-        ...
-
-
-class TransitionValidator(metaclass=SingletonMeta):
-    @classmethod
-    def validate_transition(
-        cls, transition: StatusTransitionProtocol, setting: ExpectednessSetting
-    ) -> bool:
-        E = ExpectednessSetting
-
-        # If setting is a requirement, check if it is met.
-        if setting is E.REQUIRE_NEXT_UP:
-            return transition.is_next_up
-        if setting is E.REQUIRE_FORSEEN:
-            return transition.is_foreseen
-
-        # If transition has a dangerous property, check if has been allowed.
-        if transition.is_abnormal:
-            return setting is E.ALLOW_ABNORMAL
-        if transition.is_unexpected:
-            return setting in {E.ALLOW_ABNORMAL, E.ALLOW_UNEXPECTED}
-
-        # If transition has no dangerous properties, and setting is no requirement.
-        return True
 
 
 class UpdateOrderStatusService(UpdateOrderStatusAPI):
