+++ tests/test_domain/test_services/test_update_order_status_service.py
@@ -2,10 +2,8 @@ from domain.models.identifier import Identifier
 from domain.models.order import PersistedOrder
 from domain.models.order_status import Status
 from domain.models.event import DispatchableEvent
-from domain.ports.api.update_order_status_api import ExpectednessSetting
 from domain.services.update_order_status_service import (
     UpdateOrderStatusService,
-    TransitionValidator,
 )
 from domain.errors import InvalidOrderIdError, InsufficientExpectednessError
 from test_domain.dummies import (
@@ -14,57 +12,10 @@ from test_domain.dummies import (
     EventDispatcherDummy,
     StatusToEventMapperDummy,
     TransitionValidatorDummy,
-    TransitionDummy,
 )
 from typing import Callable
 from dataclasses import dataclass
 import pytest
-
-
-@pytest.fixture
-def transition_dummy() -> TransitionDummy:
-    return TransitionDummy(from_status=Status.PENDING, to_status=Status.PENDING)
-
-
-def test_transition_validator(transition_dummy: TransitionDummy) -> None:
-    E = ExpectednessSetting
-
-    def is_valid(setting: E):
-        return TransitionValidator.validate_transition(
-            transition=transition_dummy, setting=setting
-        )
-
-    # validate with all 4 available settings for each state of the dummy.
-
-    transition_dummy.set_abnormal()
-    assert is_valid(E.ALLOW_ABNORMAL) is True
-    assert is_valid(E.ALLOW_UNEXPECTED) is False
-    assert is_valid(E.REQUIRE_FORSEEN) is False
-    assert is_valid(E.REQUIRE_NEXT_UP) is False
-
-    transition_dummy.set_unexpected()
-    assert is_valid(E.ALLOW_ABNORMAL) is True
-    assert is_valid(E.ALLOW_UNEXPECTED) is True
-    assert is_valid(E.REQUIRE_FORSEEN) is False
-    assert is_valid(E.REQUIRE_NEXT_UP) is False
-
-    transition_dummy.set_foreseen()
-    assert is_valid(E.ALLOW_ABNORMAL) is True
-    assert is_valid(E.ALLOW_UNEXPECTED) is True
-    assert is_valid(E.REQUIRE_FORSEEN) is True
-    assert is_valid(E.REQUIRE_NEXT_UP) is False
-
-    transition_dummy.set_next_up()
-    assert is_valid(E.ALLOW_ABNORMAL) is True
-    assert is_valid(E.ALLOW_UNEXPECTED) is True
-    assert is_valid(E.REQUIRE_FORSEEN) is True
-    assert is_valid(E.REQUIRE_NEXT_UP) is True
-
-
-def test_transition_validator_is_singleton() -> None:
-    validator1 = TransitionValidator()
-    validator2 = TransitionValidator()
-    assert validator1 is validator2
 
 
 @dataclass
